+++ nodeconductor/structure/serializers.py
@@ -350,7 +350,7 @@ class NestedProjectPermissionSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = models.ProjectPermission
-        fields = ['url', 'uuid', 'name', 'role', 'permission']
+        fields = ['url', 'uuid', 'name', 'role', 'permission', 'expiration_time']
 
 
 class CustomerUserSerializer(serializers.ModelSerializer):
@@ -474,9 +474,11 @@ class CustomerPermissionSerializer(PermissionFieldFilteringMixin, BasePermission
         customer = validated_data['customer']
         user = validated_data['user']
         role = validated_data['role']
-        created_by = self.context['request'].user
+        expiration_time = validated_data.get('expiration_time')
 
-        permission, _ = customer.add_user(user, role, created_by)
+        created_by = self.context['request'].user
+        permission, _ = customer.add_user(user, role, created_by, expiration_time)
+
         return permission
 
     def validate_expiration_time(self, value):
@@ -541,9 +543,10 @@ class ProjectPermissionSerializer(PermissionFieldFilteringMixin, BasePermissionS
         project = validated_data['project']
         user = validated_data['user']
         role = validated_data['role']
-        created_by = self.context['request'].user
+        expiration_time = validated_data.get('expiration_time')
 
-        permission, _ = project.add_user(user, role, created_by)
+        created_by = self.context['request'].user
+        permission, _ = project.add_user(user, role, created_by, expiration_time)
 
         return permission
 
@@ -1118,22 +1121,6 @@ class ManagedResourceSerializer(BasicResourceSerializer):
     customer_name = serializers.ReadOnlyField(source='service_project_link.project.customer.name')
 
 
-class RelatedResourceSerializer(BasicResourceSerializer):
-    url = serializers.SerializerMethodField()
-    service_tags = serializers.SerializerMethodField()
-
-    def get_url(self, resource):
-        try:
-            return reverse(resource.get_url_name() + '-detail',
-                           kwargs={'uuid': resource.uuid.hex},
-                           request=self.context['request'])
-        except NoReverseMatch:
-            return None
-
-    def get_service_tags(self, resource):
-        return resource.service_project_link.service.settings.get_tags()
-
-
 class BaseResourceSerializer(six.with_metaclass(ResourceSerializerMetaclass,
                              core_serializers.RestrictedSerializerMixin,
                              MonitoringSerializerMixin,
@@ -1191,7 +1178,6 @@ class BaseResourceSerializer(six.with_metaclass(ResourceSerializerMetaclass,
 
     tags = serializers.ReadOnlyField(source='get_tags')
     access_url = serializers.SerializerMethodField()
-    related_resources = RelatedResourceSerializer(source='get_related_resources', many=True, read_only=True)
 
     class Meta(object):
         model = NotImplemented
@@ -1204,7 +1190,7 @@ class BaseResourceSerializer(six.with_metaclass(ResourceSerializerMetaclass,
             'customer', 'customer_name', 'customer_native_name', 'customer_abbreviation',
             'tags', 'error_message',
             'resource_type', 'state', 'created', 'service_project_link', 'backend_id',
-            'access_url', 'related_resources'
+            'access_url'
         )
         protected_fields = ('service', 'service_project_link')
         read_only_fields = ('start_time', 'error_message', 'backend_id')